% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounceMarker.R
\name{addBounceMarkers}
\alias{addBounceMarkers}
\title{Add Bounce Markers to map}
\usage{
addBounceMarkers(map, lat, lng)
}
\arguments{
\item{map}{map object created by [leaflet::leaflet]}

\item{lat}{numeric latitude}

\item{lng}{numeric longitude}
}
\description{
Add Bounce Markers to map
}
\examples{
# leaflet() \%>\% addTiles() \%>\% addBounceMarkers(49, 11)
}
