% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletSidebar.R
\name{addLeafletSidebar}
\alias{addLeafletSidebar}
\alias{removeLeafletSidebar}
\alias{openLeafletSidebar}
\alias{closeLeafletSidebar}
\title{Add responsive sidebar to a leaflet map}
\usage{
addLeafletSidebar(map, id, options = NULL)

removeLeafletSidebar(map)

openLeafletSidebar(map, tab_id)

closeLeafletSidebar(map)
}
\arguments{
\item{map}{a map widget object}

\item{id}{character id}

\item{options}{list of options, e.g. \code{options = list(position = "right")}}
}
\value{
modified map
}
\description{
Types of events: Opening, closing and content
}
\examples{
\dontrun{
library(leaflet)
pane1 <- ' <h1 class="sidebar-header"> Tab1
<span class="sidebar-close"><i class="fa fa-caret-left"></i></span></h1>
<p>A responsive sidebar for <a href="http://leafletjs.com/">Leaflet</a>.</p>
'
pane2 <- '<h1 class="sidebar-header">Tab2<span class="sidebar-close"><i class="fa fa-caret-left"></i></span></h1>'
leaflet() \%>\% addTiles() \%>\% addLeafletSidebar(id = "a", options = list(position = "right")) \%>\% 
  addPanel(sidebarId = "a", layerId = "Tab1", tab_name = "Tab1", pane = pane1)

}

library(leaflet)
leaflet() \%>\% addTiles() \%>\% addLeafletSidebar(id = "a", options = list(position = "right")) \%>\% 
  removeLeafletSidebar()
}
\seealso{
\href{https://github.com/nickpeihl/leaflet-sidebar-v2}{Github: leaflet-sidebar}
}
