% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMousePosition.R
\name{addMousePosition}
\alias{addMousePosition}
\title{Add Mouse Position to map}
\usage{
addMousePosition(map)
}
\arguments{
\item{map}{map object created by \link[leaflet:leaflet]{leaflet::leaflet}}
}
\description{
Add coordinates of current mouse position to map.
}
\examples{
library(leaflet)
# leaflet() \%>\% addTiles() \%>\% addMousePosition()
}
\seealso{
\href{https://github.com/ardhi/Leaflet.MousePosition}{Github: Leaflet.MousePosition}
}
