% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletSidebar.R
\name{addPanel}
\alias{addPanel}
\title{Add panel to leaflet sidebar}
\usage{
addPanel(map, layerId, tab_name, pane, position = "top")
}
\arguments{
\item{map}{a map widget object}

\item{tab_name}{character the name of the tab}

\item{position}{character: top or bottom}
}
\value{
modified map
}
\description{
Add panel to leaflet sidebar
}
