% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletSidebar.R
\name{addResponsiveSidebar}
\alias{addResponsiveSidebar}
\title{Add responsive sidebar to the map}
\usage{
addResponsiveSidebar(map)
}
\arguments{
\item{map}{a map widget object}
}
\value{
modified map
}
\description{
Add responsive sidebar to the map
}
\examples{
library(leaflet)
leaflet() \%>\% addTiles() \%>\% addResponsiveSidebar()
}
\seealso{
\href{https://github.com/nickpeihl/leaflet-sidebar-v2}{Github: leaflet-sidebar}
}
