% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletSidebar.R
\name{addLeafletSidebar}
\alias{addLeafletSidebar}
\title{Add responsive sidebar to a leaflet map}
\usage{
addLeafletSidebar(map, options)
}
\arguments{
\item{map}{a map widget object}

\item{options}{list of options, e.g. \code{options = list(position = "right")}}
}
\value{
modified map
}
\description{
Types of events: Opening, closing and content
}
\examples{
\dontrun{
library(leaflet)
pane1 <- ' <h1 class="sidebar-header"> Tab1
<span class="sidebar-close"><i class="fa fa-caret-left"></i></span></h1>
<p>A responsive sidebar for <a href="http://leafletjs.com/">Leaflet</a>.</p>
'
pane2 <- '<h1 class="sidebar-header">Tab2<span class="sidebar-close"><i class="fa fa-caret-left"></i></span></h1>'
leaflet() \%>\% addTiles() \%>\% addLeafletSidebar(options = list(position = "right")) \%>\% 
  addPanel(layerId = "Tab1", tab_name = "Tab1", pane = pane)

}

}
\seealso{
\href{https://github.com/nickpeihl/leaflet-sidebar-v2}{Github: leaflet-sidebar}
}
